i18n support for SLS definitions. Details follow.
We store both the template (.xml.in) and the i18n-merged definition (.xml)
to make things easier for packagers and setup scripts.
POTFILES.in is generated based on files in sls/*.xml.in.
translations.sh will update sls-default.pot, merge individual translations,
and generate a clean, i18n-ized SLS definition back in the sls/ directory.

We still don't support _loading_ translations in the daemon, yet, but
this will be implemented very soon.